pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "oceanus-env"
version = "1.0.0"
description = "Multi-layer ghost-gear recovery and treaty negotiation arena for LLM RL training"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aakarsh", email = "aakarsh2007@example.com"}
]
keywords = ["reinforcement-learning", "multi-agent", "openenv", "llm", "ocean"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openenv-core>=0.1.0",
    "numpy>=1.24.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
training = [
    "transformers>=4.39.0",
    "trl>=0.8.0",
    "peft>=0.10.0",
    "accelerate>=0.28.0",
    "datasets>=2.18.0",
    "huggingface-hub>=0.21.0",
    "matplotlib>=3.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/Aakarsh2007/Oceanus-"
Space = "https://aakarsh2007-oceanus-ai.hf.space"
Notebook = "https://colab.research.google.com/drive/1lJ_EKDEZptQorExyGS6jiMMX98Qw-uOs"

[tool.setuptools.packages.find]
where = ["."]
include = ["oceanus*", "dashboard*", "train*", "server*"]
